Refactor: 불필요한 테이블과 칼럼 제거 및 TopicImage 테이블 생성해라 - #10
