--- talkPick-character/build.gradle
@@ -1,13 +1,0 @@
-bootJar.enabled = true
-
-jar.enabled = true
-
-dependencies {
-    // JPA (postgresql) 관련 의존성
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-    implementation 'org.springframework.boot:spring-boot-starter-jdbc'
-    implementation 'org.postgresql:postgresql:42.7.1'
-
-    // common
-    implementation project(':talkPick-common')
-}
