--- talkPick-character/src/main/java/talkPick/CharacterApplication.java
@@ -1,11 +1,0 @@
-package talkPick;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CharacterApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(CharacterApplication.class, args);
-    }
-}
