--- talkPick-character/src/main/java/talkPick/domain/type/Keyword.java
@@ -1,4 +1,0 @@
-package talkPick.domain.type;
-
-public enum Keyword {
-}
